Factor out hooks for orphan logs waterfall and add timeline to debug
