+++ frontend/src/pages/RequestDetailsPage/RequestDetailsPageV2/Otel.tsx
@@ -1,22 +1,10 @@
-import { MizuOrphanLog, isMizuOrphanLog } from "@/queries";
 import { useOtelTraces } from "@/queries";
-import { OtelEvent, useOtelTrace } from "@/queries/traces-otel";
-import { useMemo } from "react";
+import { useOtelTrace } from "@/queries/traces-otel";
 import { EmptyState } from "../EmptyState";
 import { SkeletonLoader } from "../SkeletonLoader";
 import { usePagination } from "../hooks";
-import { getString } from "../v2/otel-helpers";
 import { RequestDetailsPageContentV2 } from "./RequestDetailsPageV2Content";
-
-const safeParseJson = (jsonString: string) => {
-  try {
-    const parsed = JSON.parse(jsonString);
-    return parsed;
-  } catch (error) {
-    console.error("Failed to parse JSON:", error);
-    return jsonString;
-  }
-};
+import { useOrphanLogs } from "./useOrphanLogs";
 
 export function Otel({
   traceId,
@@ -46,33 +34,7 @@ export function Otel({
   });
 
   // NOTE - Flatten out events into orphan logs to allow the UI to render them
-  const orphanLogs = useMemo(() => {
-    const orphans: MizuOrphanLog[] = [];
-    for (const span of spans ?? []) {
-      if (span.events) {
-        for (const event of span.events) {
-          // TODO - Visualize other types of events on the timeline?
-          if (event.name === "log") {
-            let args =
-              safeParseJson(getString(event.attributes.arguments)) || [];
-            if (!Array.isArray(args)) {
-              args = [args];
-            }
-            // TODO - Use a more deterministic ID - preferably string that includes the trace+span+event_index
-            const logId = Math.floor(Math.random() * 1000000);
-            const orphanLog = convertEventToOrphanLog(traceId, logId, event);
-            // HACK - We want to be sure that we construct a valid orphan log, otherwise the UI will break
-            if (isMizuOrphanLog(orphanLog)) {
-              orphans.push(orphanLog);
-            } else {
-              console.error("Constructed invalid orphan log", orphanLog);
-            }
-          }
-        }
-      }
-    }
-    return orphans;
-  }, [spans, traceId]);
+  const orphanLogs = useOrphanLogs(traceId, spans ?? []);
 
   if (error) {
     console.error("Error!", error);
@@ -93,27 +55,4 @@ export function Otel({
       pagination={pagination}
     />
   );
-}
-
-/**
- * Converts an Otel event to a so-called Orphan Log to maintain backwards compatibility with the old Mizu data format
- */
-function convertEventToOrphanLog(
-  traceId: string,
-  logId: number,
-  event: OtelEvent,
-) {
-  const argsAsString = getString(event.attributes.arguments);
-  const parsedArgs = argsAsString ? safeParseJson(argsAsString) : [];
-
-  return {
-    id: logId,
-    traceId,
-    level: getString(event.attributes.level),
-    args: parsedArgs || [],
-    timestamp: event.timestamp,
-    message: getString(event.attributes.message),
-    createdAt: event.timestamp,
-    updatedAt: event.timestamp,
-  };
 }
