+++ frontend/src/pages/RequestDetailsPage/RequestDetailsPageV2/RequestDetailsPageV2Content.tsx
@@ -20,6 +20,7 @@ import { EmptyState } from "../EmptyState";
 import { TraceDetailsTimeline, TraceDetailsV2 } from "../v2";
 import { HttpSummary, SummaryV2 } from "../v2/SummaryV2";
 import { getVendorInfo } from "../v2/vendorify-traces";
+import { useRequestWaterfall } from "./useRequestWaterfall";
 
 export type SpanWithVendorInfo = {
   span: OtelSpan;
@@ -44,38 +45,7 @@ export function RequestDetailsPageContentV2({
     handleNextTrace: () => void;
   };
 }) {
-  const spansWithVendorInfo: Array<SpanWithVendorInfo> = useMemo(
-    () =>
-      spans.map((span) => ({
-        span,
-        vendorInfo: getVendorInfo(span),
-      })),
-    [spans],
-  );
-
-  // HACK - normally we'd look for the root span by trying to find the span with the parent_span_id === null
-  //        but we set a fake parent_span_id for the root span in the middleware for now
-  const rootSpan = spansWithVendorInfo.find(
-    // (item) => item.span.parent_span_id === null,
-    (item) => item.span.name === "request",
-  );
-
-  const waterfall = useMemo((): Waterfall => {
-    return [...spansWithVendorInfo, ...orphanLogs].sort((a, b) => {
-      const timeA = "span" in a ? a.span.start_time : a.timestamp;
-      const timeB = "span" in b ? b.span.start_time : b.timestamp;
-      if (timeA === timeB) {
-        // If the times are the same, we need to sort giving the priority to the root span
-        if ("span" in a && a?.span?.name === "request") {
-          return -1;
-        }
-        if ("span" in b && b?.span?.name === "request") {
-          return 1;
-        }
-      }
-      return new Date(timeA).getTime() - new Date(timeB).getTime();
-    });
-  }, [spansWithVendorInfo, orphanLogs]);
+  const { rootSpan, waterfall } = useRequestWaterfall(spans, orphanLogs);
 
   if (!rootSpan) {
     return <EmptyState />;
