+++ frontend/src/utils/index.ts
@@ -160,3 +160,13 @@ export function formatHeaders(headers: Record<string, string>): string {
     .map(([key, value]) => `${key}: ${value}`)
     .join("\n");
 }
+
+export const safeParseJson = (jsonString: string) => {
+  try {
+    const parsed = JSON.parse(jsonString);
+    return parsed;
+  } catch (error) {
+    console.error("Failed to parse JSON:", error);
+    return jsonString;
+  }
+};
